Fixed name and route of JWT file
